Removed '--recycle' flag from 'regolith' command, added '--path-states' to 'regolith clean' command (removes the dir_hash_pairs.json file from cache). Recycled copy is now the default behavior. Added deprecation comment to some of the functions (they should be removed later). Updated tests to match new behavior. Failing during operations that use recycled copies clears the dir_hash_pairs.json file (the file with the states of the file paths).

+++ regolith/main_functions.go
@@ -161,9 +161,6 @@ func UpdateAll(debug bool) error {
 // should be printed.
 func Run(profile string, debug bool) error {
 	InitLogging(debug)
-	if ExperimentalFeatureRecycleFiles {
-		Logger.Warn("The \"recycle\" feature is experimental and may be unstable.")
-	}
 	if profile == "" {
 		profile = "dev"
 	}
@@ -246,16 +243,23 @@ func Init(debug bool) error {
 //
 // The "debug" parameter is a boolean that determines if the debug messages
 // should be printed.
-func Clean(debug bool) error {
+func Clean(debug bool, cachedStatesOnly bool) error {
 	InitLogging(debug)
 	Logger.Infof("Cleaning cache...")
-	err := os.RemoveAll(".regolith")
-	if err != nil {
-		return WrapError(err, "failed to remove .regolith folder")
-	}
-	err = os.Mkdir(".regolith", 0666)
-	if err != nil {
-		return WrapError(err, "failed to recreate .regolith folder")
+	if cachedStatesOnly {
+		err := ClearCachedStates()
+		if err != nil {
+			return WrapError(err, "Failed to remove cached path states.")
+		}
+	} else {
+		err := os.RemoveAll(".regolith")
+		if err != nil {
+			return WrapError(err, "failed to remove .regolith folder")
+		}
+		err = os.Mkdir(".regolith", 0666)
+		if err != nil {
+			return WrapError(err, "failed to recreate .regolith folder")
+		}
 	}
 	Logger.Infof("Cache cleaned.")
 	return nil
